+++ ts/domlib/BUILD.bazel
@@ -22,3 +22,8 @@ typescript(
 prettier_test()
 
 eslint_test()
+
+jest_test(
+    env = "jsdom",
+    deps = [],
+)
