+++ ts/domlib/index.ts
@@ -2,3 +2,4 @@
 // License: GNU AGPL, version 3 or later; http://www.gnu.org/licenses/agpl.html
 
 export * as location from "./location";
+export * as surround from "./surround";
