+++ ts/domlib/surround/child-node-range.ts
@@ -1,8 +1,8 @@
 // Copyright: Ankitects Pty Ltd and contributors
 // License: GNU AGPL, version 3 or later; http://www.gnu.org/licenses/agpl.html
 
-import { ascend } from "../node";
-import { nodeIsElement, elementIsEmpty } from "../dom";
+import { ascend } from "../../lib/node";
+import { nodeIsElement, elementIsEmpty } from "../../lib/dom";
 
 export interface ChildNodeRange {
     parent: Node;
