+++ ts/domlib/surround/merge-match.ts
