+++ ts/domlib/surround/normalize-insertion-ranges.ts
