+++ ts/domlib/surround/text-node.ts
@@ -1,8 +1,8 @@
 // Copyright: Ankitects Pty Ltd and contributors
 // License: GNU AGPL, version 3 or later; http://www.gnu.org/licenses/agpl.html
 
+import { nodeIsElement, nodeIsText } from "../../lib/dom";
 import { Position } from "../location";
-import { nodeIsElement, nodeIsText } from "../dom";
 
 /**
  * @returns Split text node to end direction
