+++ ts/editor/index.ts
@@ -10,7 +10,7 @@ import "./editor-base.css";
 
 import "../sveltelib/export-runtime";
 import "../lib/register-package";
-import "../lib/surround";
+import "../domlib/surround";
 
 import { filterHTML } from "../html-filter";
 import { execCommand } from "./helpers";
