--- ts/lib/location/index.ts
@@ -1,2 +1,0 @@
-export { getNodeCoordinates, findNodeFromCoordinates } from "./node";
-export { Position } from "./location";
